ogma-core: Move diagram processing functions into new internal module. Refs #386.
The `overview` command is capable of processing properties or requirement
and giving information about them, but cannot process diagrams. To make
Ogma useful for all MBSE information needed for a project, it is
important that all commands be able to process any input files that Ogma
can process and that may be expected in a mission. Ogma's `overview`
command should be able to handle state machine diagrams, just like the
`diagram` command does.

To make the `overview` command handle diagrams and, consequently,
`ogma-core`'s `Command.Overview` module do the same, we need to make
common functionality in the `Command.Diagram` module available to other
commands, to reduce duplication without exposing internal types and
functions in the API.

This commit introduces a new module, `Command.CommonDiagram` that
contains definitions extracted from the `Command.Diagram` module. The
new module contains the definition of a diagram, the diagram format and
functions to parse diagrams. A new function, `readDiagram`, implements
functionality needed to read a diagram from a file, reducing the
complexity of `Command.Diagram`.
